Add 手动 Link variety和subStage表
执行在Databasedef下子文件夹内的sql文件

+++ DatatableDef/SubStage.sql
@@ -9,7 +9,7 @@ CREATE TABLE subStage (
   reqAvgTemp INT(10),
   maxFieldHumidity INT(10),
   minFieldHumidity INT(10),
-  durationLow DOUBLE(10),
-  durationMid DOUBLE(10),
-  durationHigh DOUBLE(10)
+  durationLow DOUBLE,
+  durationMid DOUBLE,
+  durationHigh DOUBLE
 ) ENGINE=InnoDB AUTO_INCREMENT=10 DEFAULT CHARSET=utf8;
